Address PR comment
